requirements.txt
playwright
fastapi
uvicorn
pillow